+++ src/core/mod.rs
@@ -7,6 +7,10 @@ mod macros;
 
 mod arc;
 mod object_type;
+mod os_err;
+mod os_status;
 
 pub use arc::*;
 pub use object_type::*;
+pub use os_err::*;
+pub use os_status::*;